keep it spicy with a what-if or two
This photo of a car that is a Jalopy but may not be the model/year expected is Creative Commons and made me chuckle: https://en.wikipedia.org/wiki/Decrepit_car#/media/File:Joshua_Tree_-_Love_car.jpg   Some "What If" thought experiements could be of value to fill out the major gaps in the main idea, so long as it does not risk confusing people. Perhaps after ensuring the intended conclusions are unmistakably unavoidable we could toss a "what if the facts were different in this one way" legal grenade.  Definitely not in the fashion of a "gotcha!" or "devil's advocate" blow hard lawyer but could a basic fraud factor be used show how sometimes the buyer is the problem and sometimes the seller is the problem and sometimes the lender is the problem and in all cases the Blockchain is flexible enough to be fit for purpose over time to avoid the traps all together.  If the car "as delivered" was a lemon or the dealer forged a signature we could, for instance, assert that the code could be configured to trigger an online mediator at a certain point who could ask some questions, get some evidence and quickly identify whether some crime, fraud or otherwise non-obvious hucksterism is in play.